feat: add table user and refresh token jwt zod and validations schemas auth and att swagger and readme

+++ src/services/delete-notice-service.ts
@@ -1,4 +1,5 @@
 import { NewsRepository } from '../repositories/news-repository'
+
 export class DeleteNoticeService {
   private readonly newsRepository: NewsRepository
 
@@ -6,10 +7,14 @@ export class DeleteNoticeService {
     this.newsRepository = new NewsRepository()
   }
 
-  public async execute(id: number): Promise<boolean> {
+  public async execute(id: string, userId: string): Promise<boolean> {
     const notice = await this.newsRepository.getById(id)
     if (!notice) {
       return false
+    }
+
+    if (notice.userId !== userId) {
+      throw new Error('Unauthorized')
     }
 
     await this.newsRepository.delete(id)
